+++ src/math/explog_macros.h
@@ -45,16 +45,19 @@
 #define ARMA_MY_EXP custom_arma_exp
 #define ARMA_MY_LOG custom_arma_log
 
+// log1p has no OpenLibM port shipped here yet; keep std::log1p universally.
 // TODO: add and use these
 // #define MY_EXPM1 std::expm1
-// #define MY_LOG1P std::log1p
 // #define MY_LOG1PEXP
+
+#define MY_LOG1P std::log1p
 
 
 #else
 
 #define MY_EXP std::exp
 #define MY_LOG std::log
+#define MY_LOG1P std::log1p
 #define ARMA_MY_EXP arma::exp
 #define ARMA_MY_LOG arma::log